% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_normalize.r
\name{normalize}
\alias{normalize}
\title{Normalize function}
\usage{
normalize(x, method)
}
\arguments{
\item{x}{a vector of numeric values}

\item{method}{different methods to normalize the vector}
}
\value{
a normalized vector
}
\description{
Normalize function
}
