% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figure2.R
\name{plot_figure_2}
\alias{plot_figure_2}
\title{Reproduce Figure 2 (note that the reproduced figure lacks annotations
in the main text; these annotations were added via the software Illustrator)}
\usage{
plot_figure_2()
}
\description{
Reproduce Figure 2 (note that the reproduced figure lacks annotations
in the main text; these annotations were added via the software Illustrator)
}
