% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figure3.R
\name{plot_figure_3}
\alias{plot_figure_3}
\title{Reproduce Figure 3 (note that the reproduced figure lacks annotations
in the main text; these annotations were added via the software Illustrator)}
\usage{
plot_figure_3()
}
\description{
Reproduce Figure 3 (note that the reproduced figure lacks annotations
in the main text; these annotations were added via the software Illustrator)
}
