% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figure4.R
\name{plot_figure_4}
\alias{plot_figure_4}
\title{Reproduce Figure 4 (note that the reproduced figure lacks annotations
in the main text; these annotations were added via the software Illustrator)}
\usage{
plot_figure_4()
}
\description{
Reproduce Figure 4 (note that the reproduced figure lacks annotations
in the main text; these annotations were added via the software Illustrator)
}
