% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appendix_E.R
\name{plot_figure_S11}
\alias{plot_figure_S11}
\title{Reproduce Figure S11 (note that the reproduced figure lacks annotations
in the main text; these annotations were added via the software Illustrator)}
\usage{
plot_figure_S11()
}
\description{
Reproduce Figure S11 (note that the reproduced figure lacks annotations
in the main text; these annotations were added via the software Illustrator)
}
